LayoutTests/pointer-lock/mouse-event-api-expected.txt

PASS mouseEvent.webkitMovementX is defined.
PASS mouseEvent.webkitMovementY is defined.
PASS successfullyParsed is true

TEST COMPLETE
